requirements.txt
flask==3.1.2
flask-restx==1.3.2
flask-limiter==4.0.0
flask-cors==6.0.1
instaloader==4.15
python-dotenv==1.2.1
validators==0.35.0
bleach==6.3.0
requests==2.32.5
TikTokApi==7.2.1